+++ coroutines/src/jsTest/kotlin/SortedBinaryTreeNodeTests.js.kt
@@ -1,0 +1,2 @@
+actual val AllowDeepInsertOnWorksTest: Boolean
+    get() = false